feat: moves layout code to newly created lib (#27)
