+++ apps/client/src/app/app.component.html
@@ -1,23 +1,3 @@
-<mat-toolbar color="primary">
-  <button
-    mat-icon-button
-    data-test-id="navigtationButton"
-    aria-label="Open Navigation"
-    (click)="onToggleNavigation()"
-  >
-    <mat-icon>menu</mat-icon>
-  </button>
-  <span class="mr-4">{{ title }}</span>
-</mat-toolbar>
-<main>
-  <mat-drawer-container>
-    <mat-drawer
-      data-test-id="sidebar"
-      mode="side"
-      [opened]="isSidebarNavigationOpen"
-    ></mat-drawer>
-    <mat-drawer-content>
-      <router-outlet></router-outlet>
-    </mat-drawer-content>
-  </mat-drawer-container>
-</main>
+<my-stack-layout [title]="title">
+  <router-outlet></router-outlet>
+</my-stack-layout>
