+++ apps/client/src/app/app.component.spec.ts
@@ -1,13 +1,7 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing";
-import { By } from "@angular/platform-browser";
 import { NoopAnimationsModule } from "@angular/platform-browser/animations";
 
 import { AppComponent } from "./app.component";
-
-const CssSelectors = {
-  ToggleNavigationButton: By.css('[data-test-id="navigtationButton"]'),
-  Sidebar: By.css('[data-test-id="sidebar"]'),
-};
 
 describe("AppComponent", () => {
   let fixture: ComponentFixture<AppComponent>;
@@ -30,35 +24,5 @@ describe("AppComponent", () => {
 
   it("should match snapshot", () => {
     expect(fixture).toMatchSnapshot();
-  });
-
-  describe("onToggleNavigation", () => {
-    it("should open the navigation", () => {
-      component.toggleSidebarNavigation = false;
-
-      fixture.debugElement
-        .query(CssSelectors.ToggleNavigationButton)
-        .triggerEventHandler("click", {});
-      fixture.detectChanges();
-
-      expect(
-        fixture.debugElement.query(CssSelectors.Sidebar).componentInstance
-          .opened
-      ).toBe(true);
-    });
-
-    it("should close the navigation", () => {
-      component.toggleSidebarNavigation = true;
-
-      fixture.debugElement
-        .query(CssSelectors.ToggleNavigationButton)
-        .triggerEventHandler("click", {});
-      fixture.detectChanges();
-
-      expect(
-        fixture.debugElement.query(CssSelectors.Sidebar).componentInstance
-          .opened
-      ).toBe(false);
-    });
   });
 });
