+++ libs/shared/component-layout/src/lib/layout.component.scss
@@ -1,0 +1,9 @@
+:host {
+  mat-drawer-container {
+    height: 100vh;
+
+    mat-drawer {
+      width: 248px;
+    }
+  }
+}
