+++ libs/shared/component-layout/src/lib/layout.component.spec.ts
@@ -1,6 +1,13 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing";
+import { By } from "@angular/platform-browser";
+import { NoopAnimationsModule } from "@angular/platform-browser/animations";
 
 import { LayoutComponent } from "./layout.component";
+
+const CssSelectors = {
+  ToggleNavigationButton: By.css('[data-test-id="navigtationButton"]'),
+  Sidebar: By.css('[data-test-id="sidebar"]'),
+};
 
 describe("LayoutComponent", () => {
   let component: LayoutComponent;
@@ -8,7 +15,7 @@ describe("LayoutComponent", () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [LayoutComponent],
+      imports: [LayoutComponent, NoopAnimationsModule],
     }).compileComponents();
 
     fixture = TestBed.createComponent(LayoutComponent);
@@ -22,5 +29,35 @@ describe("LayoutComponent", () => {
 
   it("should match snapshot", () => {
     expect(fixture).toMatchSnapshot();
+  });
+
+  describe("onToggleNavigation", () => {
+    it("should open the navigation", () => {
+      component.toggleSidebarNavigation = false;
+
+      fixture.debugElement
+        .query(CssSelectors.ToggleNavigationButton)
+        .triggerEventHandler("click", {});
+      fixture.detectChanges();
+
+      expect(
+        fixture.debugElement.query(CssSelectors.Sidebar).componentInstance
+          .opened
+      ).toBe(true);
+    });
+
+    it("should close the navigation", () => {
+      component.toggleSidebarNavigation = true;
+
+      fixture.debugElement
+        .query(CssSelectors.ToggleNavigationButton)
+        .triggerEventHandler("click", {});
+      fixture.detectChanges();
+
+      expect(
+        fixture.debugElement.query(CssSelectors.Sidebar).componentInstance
+          .opened
+      ).toBe(false);
+    });
   });
 });
